analyzer/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-request-logging==0.7.5
gunicorn==20.1.0
numpy==1.23.5
Pillow==9.3.0
python-decouple==3.6
